Split MAST builder lifecycle modules
